+++ examples/shapes.rs
@@ -55,7 +55,7 @@ impl WindowBehavior for Test {
         &'pass mut self,
         mut window: Window<'_>,
         graphics: &mut kludgine::RenderingGraphics<'_, 'pass>,
-    ) -> bool {
+    ) {
         window.redraw_in(Duration::from_millis(16));
         self.angle += Angle::degrees(180) * window.elapsed();
         self.dips_square
@@ -66,6 +66,5 @@ impl WindowBehavior for Test {
             .translate_by(Point::new(BLUE_TRIANGLE_SIZE / 2, BLUE_TRIANGLE_SIZE / 2))
             .rotate_by(self.angle)
             .render(graphics);
-        true
     }
 }