+++ examples/texture-collection.rs
@@ -58,7 +58,7 @@ impl WindowBehavior for Test {
         &'pass mut self,
         mut window: Window<'_>,
         graphics: &mut kludgine::RenderingGraphics<'_, 'pass>,
-    ) -> bool {
+    ) {
         window.redraw_in(Duration::from_millis(16));
         self.angle += Angle::degrees(180) * window.elapsed();
         self.k
@@ -78,7 +78,6 @@ impl WindowBehavior for Test {
                 .cast(),
             )
             .render(graphics);
-        true
     }
 
     fn clear_color(&self) -> Option<Color> {